doDelete, Util.jsReplace

+++ src/main/webapp/WEB-INF/jsp/usr/article/detail.jsp
@@ -37,7 +37,7 @@
 				
 				<c:if test="${loginedMemberId == article.memberId }">
 					<a href="modify?id=${article.id }">수정</a>
-					<a href="doDelete?id=${article.id }">삭제</a>
+					<a href="doDelete?id=${article.id }" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a>
 				</c:if>
 			</div>
 		</div>
